Do not use Codacy.
Codacy seems to provide little value to projects already instrumented
with .NET analyzers and using GH actions.  Its most interesting tool
is SonarAnalyzer.CSharp, that is a Roslyn analyzer itself and can thus
be added directly to projects.

We will add SonarAnalyzer.CSharp some day (probably via Buildvana SDK)
and configure its rules in .globalconfig; that's going to be a rather
long and tedious process, as there's a lot of overlap between Sonar
and standard .NET rules: some are the same, some are in direct conflict,
some are the same but configured differently, etc.
For now we'd better concentrate on other stuff.

Signed-off-by: Riccardo De Agostini <me@riccar.do>