ci: Explicitly set token for actions/checkout
The README says ${{github.token}} is the default value for 'token'.
https://github.com/actions/checkout/tree/v3.1.0#usage

The source code says 'token' is a required input.
https://github.com/actions/checkout/blob/v3.1.0/src/input-helper.ts#L110

It *seems* to work (checkout succeeds) without it
(which is weird - isn't it a *required* input?)
but the repo is public, can't be sure that the persisted credentials
will be valid for pushing and creating a release.
No harm in setting it explicitly anyway.

Signed-off-by: Riccardo De Agostini <me@riccar.do>